Don't use transactions (multi) for bitops
If you use a bitop against a frequently changing bitmap,
the multi can be cancelled because the underlying key tracked
by watch will change and invalidate the transaction.  The
biota will then return invalid results (0).  This removes the
transaction block.  Unknown if this is dangerous if the bitmap is
in the process of being copied.  Causes bitops to not be strongly
consistent.